requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy==2.0.45
celery==5.6.2
redis==7.1.0
python-multipart==0.0.21
httpx==0.28.1
python-dotenv==1.2.1
aiofiles==25.1.0
langchain==1.2.8
langchain-community==0.4.1
langchain-pymupdf4llm==0.5.0
docx2txt==0.9
langchain-text-splitters==1.1.0
pymongo==4.16.0
qdrant-client==1.16.2
groq==1.0.0
numpy==2.2.3
google-genai==1.63.0
fastembed==0.5.1
